+++ Common/src/main/java/net/merchantpug/apugli/power/factory/ModifyDurabilityChangePowerFactory.java
@@ -3,9 +3,11 @@ package net.merchantpug.apugli.power.factory;
 import io.github.apace100.apoli.data.ApoliDataTypes;
 import io.github.apace100.apoli.util.Comparison;
 import io.github.apace100.calio.data.SerializableData;
+import io.github.apace100.calio.data.SerializableDataType;
 import io.github.apace100.calio.data.SerializableDataTypes;
 import net.merchantpug.apugli.data.ApugliDataTypes;
 import net.merchantpug.apugli.platform.Services;
+import net.minecraft.util.Mth;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
@@ -17,7 +19,21 @@ public interface ModifyDurabilityChangePowerFactory<P> extends ValueModifyingPow
                 .add("item_condition", Services.CONDITION.itemDataType(), null)
                 .add("comparison", ApoliDataTypes.COMPARISON, Comparison.GREATER_THAN_OR_EQUAL)
                 .add("compare_to", SerializableDataTypes.INT, Integer.MIN_VALUE)
-                .add("comparisons", ApugliDataTypes.comparisonMap(SerializableDataTypes.INT), Map.of());
+                .add("comparisons", ApugliDataTypes.comparisonMap(SerializableDataTypes.INT), Map.of())
+                .add("function", SerializableDataType.enumValue(Function.class), Function.FLOOR);
+    }
+
+    default int postFunction(P power, double value) {
+        SerializableData.Instance data = getDataFromPower(power);
+
+        int retVal;
+        switch ((Function)data.get("function")) {
+            case CEILING -> retVal = value >= 0 ? Mth.ceil(value) : Mth.floor(value);
+            case ROUND -> retVal = (int) Math.round(value);
+            default -> retVal = value >= 0 ? Mth.floor(value) : Mth.ceil(value);
+        }
+
+        return retVal;
     }
 
     default boolean doesApply(P power, Level level, ItemStack stack, int durabilityChange) {
@@ -33,6 +49,12 @@ public interface ModifyDurabilityChangePowerFactory<P> extends ValueModifyingPow
         }
 
         return ((Comparison) data.get("comparison")).compare(durabilityChange, data.getInt("compare_to"));
+    }
+
+    enum Function {
+        FLOOR,
+        ROUND,
+        CEILING
     }
 
 }